pyproject.toml
[tool.poetry]
name = "efficient-deep-learning"
packages = [{include = "src"}]
version = "0.14.11"
description = "Library for efficient deep learning research"
authors = ["Gleb Khaykin <khaykingleb@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
typer = "^0.13.1"
attrs = "^24.2.0"
loguru = "^0.7.2"
pydantic = "^2.8.2"
polars = "^1.13.1"
pandas = "^2.2.2"
pandera = {extras = ["polars"], version = "^0.20.3"}
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
torch = "^2.5.1"
torchvision = "^0.20.1"
torchmetrics = "^1.6.0"
lightning = "^2.4.0"
wandb = "^0.18.7"
einops = "^0.8.0"
#triton = { version = "^3.0.0", markers = "sys_platform == 'linux'" }
tiktoken = "^0.8.0"
pyarrow = "^18.0.0"
matplotlib = "^3.9.2"

[tool.poetry.group.audio.dependencies]
soundfile = "^0.12.1"
torchaudio = "^2.5.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
jupyterlab = "^4.3.1"
plotly = "^5.23.0"
gpustat = "^1.1.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
pytest-xdist = "^3.6.1"
pytest-cov = "^6.0.0"

[tool.ruff]
target-version = "py310"
line-length = 79
indent-width = 4

# Exclude a variety of commonly ignored directories
exclude = [
    ".dvc",
    ".github",
    ".git",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.ruff.lint]
select = [
  "RUF",   # ruff checks
  "F",     # pyflakes
  "E",     # pycodestyle
  "W",     # pycodestyle
  "C90",   # mcabbe
  "I",     # isort
  "N",     # pep8-naming
  "D",     # pydocstyle
  "UP",    # pyupgrade
  "YTT",   # flake8-2020
  "ANN",   # flake8-annotations
  "ASYNC", # flake8-async
  "S",     # flake8-bandit
  "FBT",   # flake8-boolean-trap
  "B",     # flake8-bugbear
  "A",     # flake8-builtins
  "C4",    # flake8-comprehensions
  "DTZ",   # flake8-datetimez
  "T10",   # flake8-debugger
  "EM",    # flake8-errmsg
  "ISC",   # flake8-implicit-str-concat
  "G",     # flake8-logging-format
  "LOG",   # flake8-logging
  "INP",   # flake8-no-pep420
  "PIE",   # flake8-pie
  "T20",   # flake8-print
  "PYI",   # flake8-pyi
  "PT",    # flake8-pytest-style
  "Q",     # flake8-quotes
  "RSE",   # flake8-raise
  "RET",   # flake8-return
  "SLOT",  # flake8-slot
  "SIM",   # flake8-simplify
  "TID",   # flake8-tidy-imports
  "TCH",   # flake8-type-checking
  "INT",   # flake8-gettext
  "ARG",   # flake8-unused-arguments
  "PTH",   # flake8-use-pathlib
  "ERA",   # flake8-eradicate
  "PD",    # pandas-vet
  "PGH",   # pygrep
  "PL",    # pylint
  "TRY",   # tryceratops
  "FLY",   # flynt
  "NPY",   # NumPy-specific rules
  "PERF",  # Perflint
  "FURB",  # refurb
]
ignore = [
  "ANN101",  # missing type annotation for `self` in method
  "ANN204",  # missing return type annotation for special method
  "D105",    # missing docstring in magic method
  "PLR0913", # too many arguments in function definition ({c_args} > {max_args})
  "N812",    # imports are non-lowercase
  "S310",    # audit URL open for permitted schemes
  "S311",    # standard pseudo-random generators are not suitable for cryptographic purposes
  "ISC001",  # implicit string concatenation
]
fixable = [
    "I",      # isort
    "D",      # pydocstyle
    "TCH",    # typecheking
    "F401",   # unused import
]

[tool.ruff.lint.per-file-ignores]
# "src/collections/audio/asr/data.py" = ["ARG002", "ANN401"]
"tests/*" = [
  "ANN201",  # missing return type annotation for public function
  "D100",    # missing docstring in module
  "D103",    # missing docstring in public function
  "S101",    # use of `assert` detected
  "PLR2004", # magic value used in comparison
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "*.py"
generate_report_on_test = true
addopts = "-n auto --cov=src"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
